requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
contourpy==1.2.1
coverage==7.8.0
cssselect2==0.7.0
cssutils==2.11.1
cycler==0.12.1
datasets==3.0.1
debugpy==1.8.5
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.31.0
diffvg==0.0.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.6
hydra-core==1.3.2
identify==2.6.1
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.9
loguru==0.7.2
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==2.1.0
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft==0.13.0
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.2.2
pre_commit==4.0.1
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pyrallis==0.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-fmm==2024.9.16
scikit-image==0.24.0
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
svgelements==1.9.6
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.13.2
tifffile==2024.8.10
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.20.1
torch==2.4.0
torchmetrics==1.6.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.0
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
virtualenv==20.26.6
wandb==0.18.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.0
zmq==0.0.0
